+++ src/components/Indexer/processors/BaseProcessor.ts
@@ -215,7 +215,7 @@ export abstract class BaseEventProcessor {
           { timeout: 20000 }
         )
         return nonceResponse.status === 200 && nonceResponse.data
-          ? String(parseFloat(nonceResponse.data.nonce) + 1)
+          ? String(parseInt(nonceResponse.data.nonce) + 1)
           : Date.now().toString()
       } else {
         return Date.now().toString()
